+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased (11.0.1)]
+## [Unreleased (12.0.0)]
 
 ### Added
 
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The CMake build system no longer links `dot_builtins` against libsocket.
 - The CMake build system now only compiles `dot_builtins` if building shared
   libraries (`-DBUILD_SHARED_LIBS=ON`).
+- **Breaking**: the `polygon_t.style` member is now a struct.
 
 ### Fixed
 
@@ -2623,7 +2624,7 @@ March 13, 2000: Use AM_PROG_LIBTOOL instead of AC_PROG_LIBTOOL
    in configure.in.  John Ellson <ellson@graphviz.org>
 ```
 
-[Unreleased (11.0.1)]: https://gitlab.com/graphviz/graphviz/compare/11.0.0...main
+[Unreleased (12.0.0)]: https://gitlab.com/graphviz/graphviz/compare/11.0.0...main
 [11.0.0]: https://gitlab.com/graphviz/graphviz/compare/10.0.1...11.0.0
 [10.0.1]: https://gitlab.com/graphviz/graphviz/compare/9.0.0...10.0.1
 [9.0.0]: https://gitlab.com/graphviz/graphviz/compare/8.1.0...9.0.0